summarizar-bert-master/requirements.txt
Flask
Flask-RESTful
bert-extractive-summarizer==0.4.0
uwsgi
gunicorn==19.9.0
